[LeadGenerationBundle] [Tests] fix namespace tests
