+++ src/Kunstmaan/LeadGenerationBundle/Tests/_support/_generated/UnitTesterActions.php
@@ -1,5 +1,5 @@
-<?php  //[STAMP] 78ee096a68f63bd6a3edb6e7ad138714
-namespace LeadGenerationBundle\_generated;
+<?php  //[STAMP] 7d36f422c4195a66cde62673383e2976
+namespace Kunstmaan\LeadGenerationBundle\Tests\_generated;
 
 // This class was automatically generated by build task
 // You should not change it manually as it will be overwritten on next build
