+++ src/Kunstmaan/LeadGenerationBundle/Tests/unit.suite.yml
@@ -1,0 +1,5 @@
+actor: UnitTester
+modules:
+  enabled:
+  - Asserts
+  - Kunstmaan\LeadGenerationBundle\Tests\Helper\Unit
