+++ src/Kunstmaan/LeadGenerationBundle/Tests/unit/Entity/Popup/AbstractPopupTest.php
@@ -5,9 +5,8 @@ namespace Kunstmaan\LeadGenerationBundle\Tests\Entity\Popup;
 use Doctrine\Common\Collections\ArrayCollection;
 use Kunstmaan\LeadGenerationBundle\Entity\Rule\LocaleWhitelistRule;
 use Kunstmaan\LeadGenerationBundle\Entity\Rule\UrlBlacklistRule;
+use Kunstmaan\LeadGenerationBundle\Tests\unit\Entity\Popup\Popup;
 use PHPUnit_Framework_TestCase;
-
-
 
 /**
  * Generated by PHPUnit_SkeletonGenerator on 2012-09-04 at 16:54:04.
