+++ src/Kunstmaan/LeadGenerationBundle/Tests/unit/Entity/Rule/AfterXScrollPercentRuleTest.php
