+++ src/Kunstmaan/LeadGenerationBundle/Tests/unit/Entity/Rule/AfterXSecondsRuleTest.php
